Rename and reformat RecipesSQLite.java
